Add project members' names to the documentation
In this update, we added the names of all project members to the project documentation. This information has been included in the README file to clearly identify the contributors to the project and give credit to each team member. The names of the members are as follows:

Amin Behbudov

Fares Abdulmajeed Alabdulhadi

Tahmid Wasif Zaman

Dimural Murat

This change ensures that proper attribution is provided and aligns with best practices for open-source projects, making it clear who the collaborators are.

Additionally, the formatting of the "Authors" section in the README has been updated for clarity and consistency. This change doesn't affect the functionality of the project but is important for project attribution and recognition.